Check that inter shell edges are only being moved from nodes with high degree
Previously, this step just moved any inter shell edge, regardless of the degree of the nodes involved.
This is contrary to the behavior described in the paper, so an additional check has been created.

Also, the behavior for moving an inter shell edge was actually deleting the
inter shell edge entirely, and moving it to be an intra-shell edge.  That has also
been fixed.


Former-commit-id: 7a66a1b052515f480c3bf39f0319e2f9bb06fcf0 [formerly de4358e6c92e53df4b83b5d54ea5c8c9f4cac521] [formerly de4358e6c92e53df4b83b5d54ea5c8c9f4cac521 [formerly 6f8c64151ada643eb944345e2b59216b1705c6b8]]
Former-commit-id: 7d35e87e7696277a7e17faf3dedc90190cefb777 [formerly e3ca521b704a9abf13a03c3fbcaa6dae05731f02]
Former-commit-id: 4229fc6f89f06f730abd31c61809e07a131c158d