Fix another bug in shell rewiring.
When swapping inter-shell edges, we only care that the node outside
the current shell is the one not in the neighborhood of both nodes.  The
two nodes within the shell are perfectly fine to be connected.

When swapping intra-shell edges, we need to ensure that the edge we
are going to create actually IS an intra-shell edge.  So I created
__shell_neighborhood_of_v to check for that


Former-commit-id: ef47c3b89799eb65caf6b3897a8a9f6681d0bbe6 [formerly ce6775965d93b83278dfe5f5b5d9ca4c2c7817a8] [formerly ce6775965d93b83278dfe5f5b5d9ca4c2c7817a8 [formerly 983fa9b3dbc364f44666042129e641b40eb49ea2]]
Former-commit-id: b1c4d2d1409ee680d695174f7153c77414cea37c [formerly 7d073b3bf11ff3828d8ed2fd5ed1cf55656e94e8]
Former-commit-id: 65da3c1d6d4a3801914cd095e3b81e72c74607cc