Fix two bugs related to edge generation.
One bug was that nodes could be added to higher shells during edge generation within their shells.
For example, in a 1-shell, nodes could have out-degree greater than 1, which could lead to cycles
that actually caused a 2-core to be created.  This is being limited by ensuring
that no node has out degree greater than its shell number, instead of
the degeneracy number of the graph.

Another bug is related to nodes being added to higher shells with the addition
of intra-shell edges.  If these edges are added in a specific way, nodes
can jump core levels.  We have limited the number of inter-shell
edges each node can have as a result


Former-commit-id: d301eadd8a843f077e805e6624938b5ecc312abc [formerly 6b1491ce7153e8cdf480cba5e5346026b29cdd42] [formerly 6b1491ce7153e8cdf480cba5e5346026b29cdd42 [formerly 1ccee895b92a97c2b924d65fe607b2efcdf402c1]]
Former-commit-id: c64f3daea2811ae4a297723bf0401eba5015d084 [formerly db0d1e1287cc89bad250ac967fc42816ee524333]
Former-commit-id: 8603921721498af34c91b40d97d98d91c5ff3ba4